Readying pulse_post3D for EPG movies
- Created functions pulsetitle (creates figure titles for any pulse
simulation variables) and timeslice3D (does the reshaping and trimming
of big MFiX vectors).

- Finished volume3D so it should make EPG videos. Written to work for
any number of isosurfaces and to label according to domain size (based
on user-defined values in pulse_post3D). [NEEDS TO BE TESTED ON CLUSTER]

- Lots of little updates to pulse_post3D to clean it up and make it more
general. User can now choose which variables to process (though
currently there is only a processing function for EPG) and whether or
not to suppress image displays. [NEEDS TO BE TESTED ON CLUSTER]